move contours to separate file
